pyproject.toml
[project]
name = "docling-vlm-sidecar"
version = "0.1.0"
description = "Lightweight MLX-based document conversion sidecar for Docstra"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Docstra Team" }
]
keywords = ["docling", "mlx", "vlm", "document-conversion", "macos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mlx-vlm>=0.1.0",
    "docling-core>=2.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
    "pytest>=8.0.0",
]

[project.scripts]
docling-vlm-sidecar = "sidecar_main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=6.0.0",
    "pytest>=8.0.0",
]
